+++ packages/apollo-engine-reporting/src/index.ts
@@ -1,2 +1,2 @@
 export { EngineReportingOptions, EngineReportingAgent } from './agent';
-export { EngineFederatedTracingExtension } from './federatedExtension';
+export { default as federatedPlugin } from './federatedPlugin';