+++ packages/apollo-server-core/src/utils/dispatcher.ts
@@ -1,27 +1,47 @@
-import { GraphQLRequestListener } from "apollo-server-plugin-base";
+import { AnyFunction, AnyFunctionMap } from "apollo-server-types";
 
-type AnyFunction = (...args: any[]) => any;
 type Args<F> = F extends (...args: infer A) => any ? A : never;
 type AsFunction<F> = F extends AnyFunction ? F : never;
 type UnwrapPromise<T> = T extends Promise<infer U> ? U : T;
 
 type DidEndHook<TArgs extends any[]> = (...args: TArgs) => void;
 
-export class Dispatcher<T extends GraphQLRequestListener> {
+export class Dispatcher<T extends AnyFunctionMap> {
   constructor(protected targets: T[]) {}
+
+  private callTargets<TMethodName extends keyof T>(
+    targets: T[],
+    methodName: TMethodName,
+    ...args: Args<T[TMethodName]>
+  ): ReturnType<AsFunction<T[TMethodName]>>[] {
+    return targets.map(target => {
+      const method = target[methodName];
+      if (method && typeof method === 'function') {
+        return method.apply(target, args);
+      }
+    });
+  }
 
   public async invokeHookAsync<TMethodName extends keyof T>(
     methodName: TMethodName,
     ...args: Args<T[TMethodName]>
-  ): Promise<UnwrapPromise<ReturnType<AsFunction<T[TMethodName]>>>[]> {
+  ): Promise<ReturnType<AsFunction<T[TMethodName]>>[]> {
     return await Promise.all(
-      this.targets.map(target => {
-        const method = target[methodName];
-        if (method && typeof method === 'function') {
-          return method.apply(target, args);
-        }
-      }),
-    );
+      this.callTargets(this.targets, methodName, ...args));
+  }
+
+  public invokeHookSync<TMethodName extends keyof T>(
+    methodName: TMethodName,
+    ...args: Args<T[TMethodName]>
+  ): ReturnType<AsFunction<T[TMethodName]>>[] {
+    return this.callTargets(this.targets, methodName, ...args);
+  }
+
+  public reverseInvokeHookSync<TMethodName extends keyof T>(
+    methodName: TMethodName,
+    ...args: Args<T[TMethodName]>
+  ): ReturnType<AsFunction<T[TMethodName]>>[] {
+    return this.callTargets(this.targets.reverse(), methodName, ...args);
   }
 
   public async invokeHooksUntilNonNull<TMethodName extends keyof T>(